Create security.service package and move the services and their implementation of service related to security to this new package
